Struct slice_dst::SliceWithHeader

#[repr(C)]
pub struct SliceWithHeader<Header, Item> {
    pub header: Header,
    pub slice: [Item],
    /* private fields */
}

A custom slice-based DST.

The length is stored as a usize at offset 0. This must be the length of the trailing slice of the DST.

Fields
header: Header

The included header. Does not dictate the slice length.

slice: [Item]

The included slice.

Implementations

impl<Header, Item> SliceWithHeader<Header, Item>

pub fn new<A, I>(header: Header, items: I) -> Awhere
A: AllocSliceDst<Self>,
I: IntoIterator<Item = Item>,
I::IntoIter: ExactSizeIterator,

Create a new slice/header DST in a AllocSliceDst container.

Panics

Panics if the items iterator incorrectly reports its length.

pub fn from_slice<A>(header: Header, s: &[Item]) -> Awhere
A: AllocSliceDst<Self>,
Item: Copy,

Create a new slice/header DST from a slice, in a AllocSliceDst container.